Add info display to Access Point node  = Expose mission id to client  = Reduce node screens to non-common content size  = Fix boilerplate params for node implementations  + Add seeded-random determination of per-mission institution (+ rand crate)  + Add caching to WarehouseManager (+ cached crate)  + Warehouse: Location endpoint

+++ crates/hall/src/public/core/mission.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use shared_net::Bufferable;
 use shared_net::VSizedBuffer;
 
-pub type MissionIdType = u8;
+pub type MissionIdType = u64;
 pub type MissionNodeLinkDamageType = u8;
 
 #[repr(u8)]
